% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_charts.r
\name{compare_distribution}
\alias{compare_distribution}
\title{compare_distribution}
\usage{
compare_distribution(rundata, runvar)
}
\arguments{
\item{rundata}{tbl}

\item{runvar}{chr variable denoting the different runs}
}
\value{

}
\description{
compare_distribution
}
