% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_charts.r
\name{compare_ts}
\alias{compare_ts}
\title{compare_ts}
\usage{
compare_ts(rundata, runvar)
}
\arguments{
\item{rundata}{tbl}

\item{runvar}{chr variable denoting the different runs}
}
\value{

}
\description{
compare_ts
}
