% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_charts.r
\name{diff_by_var}
\alias{diff_by_var}
\title{diff_by_var}
\usage{
diff_by_var(rundata, pm1, pm2, by)
}
\arguments{
\item{rundata}{tbl}

\item{pm1}{var First power variable}

\item{pm2}{var Second power variable}

\item{by}{var x var to plot diff against}
}
\value{

}
\description{
diff_by_var
}
