% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coggan_formulas.r
\name{intensity_factor}
\alias{intensity_factor}
\title{intensity_factor}
\usage{
intensity_factor(ride_data, ftp)
}
\arguments{
\item{ride_data}{tbl}

\item{ftp}{dbl FTP}
}
\value{

}
\description{
intensity_factor
}
