% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_power}
\alias{simulate_power}
\title{simulate_power}
\usage{
simulate_power(n_seconds, target_ap, target_np, max_power = 1000)
}
\arguments{
\item{n_seconds}{int Number of seconds to simulate}

\item{target_ap}{dbl Average power}

\item{target_np}{dbl Normalised power}
}
\value{
Vector of power values
}
\description{
Simulate power data for a given NP and AP
}
