% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_charts.r
\name{compare_distribution}
\alias{compare_distribution}
\title{compare_distribution}
\usage{
compare_distribution(rundata, pm1, pm2)
}
\arguments{
\item{rundata}{tbl}

\item{pm1}{var First power variable}

\item{pm2}{var Second power variable}
}
\value{

}
\description{
compare_distribution
}
